Switch getInitialProps()
